Move dependency_graph method into utils.py
